Remove old toolbar and move + rename delegate
